requirements.txt
fastapi[standard]>=0.110.0
sqlalchemy>=2.0.29
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0
uvicorn>=0.29.0
python-multipart>=0.0.9
google-genai>=0.3.0
python-dotenv>=1.0.1
pytest>=8.0.0
httpx>=0.27.0
